+++ src/lib.rs
@@ -69,4 +69,4 @@ pub use db::{GAFBase, GAFBaseParams};
 pub use path_index::PathIndex;
 pub use read_set::{ReadSet, AlignmentOutput};
 pub use subgraph::Subgraph;
-pub use subgraph::query::{SubgraphQuery, HaplotypeOutput};
+pub use subgraph::query::{SubgraphQuery, HaplotypeOutput, SnarlOutput};